refactor: extract rundown parsing
